+++ utility/src/main/java/de/unistuttgart/iaas/amyassist/amy/utility/rest/Method.java
@@ -21,14 +21,40 @@
  * For more information see notice.md
  */
 
-package de.unistuttgart.iaas.amyassist.amy.httpserver.rest;
+package de.unistuttgart.iaas.amyassist.amy.utility.rest;
 
 /**
- * The possible Http Methods
+ * POJO-mapping of JSON method object
  * 
  * @author Christian Bräuner
  */
-public enum MethodType {
+public class Method {
 
-	GET, POST, DELETE;
+	private String link;
+	private String description;
+	private MethodType type;
+
+	public String getLink() {
+		return link;
+	}
+
+	public void setLink(String link) {
+		this.link = link;
+	}
+
+	public String getDescription() {
+		return description;
+	}
+
+	public void setDescription(String description) {
+		this.description = description;
+	}
+
+	public MethodType getType() {
+		return type;
+	}
+
+	public void setType(MethodType type) {
+		this.type = type;
+	}
 }
