+++ utility/src/main/java/de/unistuttgart/iaas/amyassist/amy/utility/rest/ParameterType.java
@@ -21,40 +21,13 @@
  * For more information see notice.md
  */
 
-package de.unistuttgart.iaas.amyassist.amy.httpserver.rest;
+package de.unistuttgart.iaas.amyassist.amy.utility.rest;
 
 /**
- * POJO-mapping of JSON method object
+ * possible parameter types for methods
  * 
  * @author Christian Bräuner
  */
-public class Method {
-
-	private String link;
-	private String description;
-	private MethodType type;
-
-	public String getLink() {
-		return link;
-	}
-
-	public void setLink(String link) {
-		this.link = link;
-	}
-
-	public String getDescription() {
-		return description;
-	}
-
-	public void setDescription(String description) {
-		this.description = description;
-	}
-
-	public MethodType getType() {
-		return type;
-	}
-
-	public void setType(MethodType type) {
-		this.type = type;
-	}
+public enum ParameterType {
+	STRING, INTEGER, FLOAT;
 }
